requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.46
PyMySQL==1.1.2
cryptography>=42.0.0
python-dotenv==1.2.1
gunicorn==21.2.0
web3>=6.0.0
stellar-sdk==13.2.1